requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
Django==5.1.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mssql-django==1.5
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.0
tzdata==2024.1
Werkzeug==3.0.4
zope.interface==7.2
